Updated the parser of refactoring problems to extract the timestamp of refactorings.
Refactored the file system manipulation parts of the tests.

Resolved the following conflict:
	edu.illinois.codingspectator.refactoringproblems.logger/src/edu/illinois/codingspectator/refactoringproblems/logger/ProblemChanges.java

+++ edu.illinois.codingspectator.saferecorder/src/edu/illinois/codingspectator/saferecorder/SafeRecorder.java
@@ -7,8 +7,8 @@ import java.io.File;
 import java.io.IOException;
 
 import edu.illinois.codingspectator.codingtracker.helpers.Debugger;
-import edu.illinois.codingspectator.codingtracker.helpers.FileHelper;
 import edu.illinois.codingspectator.codingtracker.helpers.Messages;
+import edu.illinois.codingspectator.codingtracker.helpers.ResourceHelper;
 import edu.illinois.codingspectator.data.CodingSpectatorDataPlugin;
 
 /**
@@ -49,20 +49,23 @@ public class SafeRecorder {
 		Debugger.debug("END COMMIT");
 		File tempFile= currentRecordFile;
 		currentRecordFile= mainRecordFile;
-		String tempContent= FileHelper.readFileContent(tempFile);
-		record(tempContent);
-		tempFile.delete();
+		//Temporary file does not exist at this point if nothing is recorded to it during data uploading
+		if (tempFile.exists()) {
+			String tempContent= ResourceHelper.readFileContent(tempFile);
+			record(tempContent);
+			tempFile.delete();
+		}
 	}
 
 	public synchronized void record(CharSequence text) {
 		try {
-			FileHelper.ensureFileExists(currentRecordFile);
+			ResourceHelper.ensureFileExists(currentRecordFile);
 		} catch (IOException e) {
 			Debugger.logExceptionToErrorLog(e, Messages.Recorder_CreateRecordFileException + currentRecordFile.getName());
 		}
 		Debugger.debugFileSize("Before: ", currentRecordFile);
 		try {
-			FileHelper.writeFileContent(currentRecordFile, text, true);
+			ResourceHelper.writeFileContent(currentRecordFile, text, true);
 		} catch (IOException e) {
 			Debugger.logExceptionToErrorLog(e, Messages.Recorder_AppendRecordFileException + currentRecordFile.getName());
 		}
